// src/app/models/RequestClient.model.ts
export class RequestClientModel {
    id?:number;
    date?:string;
    comment?:string;
    endOfRent?:string;
    adviserId?: number;
    clientId?: number;
    contractId?: number;
    guarantorId?: number;
    propertyId?: number;
    requestTypeId?: number;
    requestStatusId?: number;
    client?:any;
    property?:any;
}